Fixing focus-bug w/modal and action button
