--- components/groupDetails/groupDetailsEditButton/index.jsx
@@ -1,47 +1,0 @@
-/**
- * Copyright 2020 Inrupt Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- * Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-/* eslint react/forbid-prop-types: off, react/jsx-props-no-spreading: off */
-
-import React from "react";
-import { Button } from "@inrupt/prism-react-components";
-import GroupDetailsModal from "../groupDetailsModal";
-
-export const TESTCAFE_ID_GROUP_DETAILS_EDIT_BUTTON =
-  "group-details-edit-button";
-
-export default function GroupDetailsEditButton(props) {
-  const [open, setOpen] = React.useState(false);
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
-  return (
-    <>
-      <Button
-        onClick={handleOpen}
-        data-testid={TESTCAFE_ID_GROUP_DETAILS_EDIT_BUTTON}
-        {...props}
-      >
-        Edit Group Details
-      </Button>
-      <GroupDetailsModal handleClose={handleClose} open={open} />
-    </>
-  );
-}
